🐛 : 액세스 토큰 만료 체크가 페이지 로드 후 한번만 이뤄지는 문제 해결

+++ src/hooks/useUserSession.js
@@ -1,6 +1,7 @@
 import { ACCESS_TOKEN_COOKIE_NAME } from 'constants';
 import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import { useLocation } from 'react-router-dom';
 
 import userActions from 'store/user/action';
 import userThunk from 'store/user/thunk';
@@ -8,15 +9,20 @@ import userThunk from 'store/user/thunk';
 import { getCookie } from 'lib/cookieUtils';
 
 function useUserSession() {
+  const isLogin = useSelector(({ user }) => user.isLogin);
   const dispatch = useDispatch();
 
   useEffect(() => {
     dispatch(userThunk.getUserProfile());
   }, []);
 
-  if (!getCookie(ACCESS_TOKEN_COOKIE_NAME)) {
-    dispatch(userActions.removeInfo());
-  }
+  const { pathname } = useLocation();
+
+  useEffect(() => {
+    if (isLogin === true && !getCookie(ACCESS_TOKEN_COOKIE_NAME)) {
+      dispatch(userActions.removeInfo());
+    }
+  }, [isLogin, pathname]);
 }
 
 export default useUserSession;
